Separate gravity logic from player class
