+++ Input/Test.tscn
@@ -39,7 +39,7 @@ transform = Transform3D(-0.866023, -0.433016, 0.250001, 0, 0.499998, 0.866027, -
 shadow_enabled = true
 
 [node name="Player" parent="." instance=ExtResource("1_n7co6")]
-transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 10.2937, 139.936, 42.2251)
+transform = Transform3D(1, 0, 0, 0, 1, 0, 0, 0, 1, 10.2937, 66.936, 42.2251)
 
 [node name="InputManager" parent="." instance=ExtResource("2_5uwd3")]
